chore(Example): Move 'new' examples into examples folder
