+++ examples/collection.invalid.yaml
