+++ examples/collection.yaml
