+++ examples/datatable.ts
@@ -1,4 +1,4 @@
-import { Datatable } from '../types'
+import { Datatable } from '../ts/types'
 
 export const minimal: Datatable = {
   type: 'Datatable',
