+++ examples/mediaObject.yaml
